--- a/qiskit/backends/local/_simulatortools.py
+++ b/qiskit/backends/local/_simulatortools.py
@@ -14,6 +14,8 @@ Functions
     and b2 as the i2th bit
 
 """
+
+from string import ascii_uppercase, ascii_lowercase
 import numpy as np
 from sympy import Matrix, pi, E, I, cos, sin, N, sympify
 
@@ -99,6 +101,62 @@ def single_gate_matrix(gate, params=None):
                       -np.exp(1j*lam)*np.sin(theta/2)],
                      [np.exp(1j*phi)*np.sin(theta/2),
                       np.exp(1j*phi+1j*lam)*np.cos(theta/2)]])
+
+
+def einsum_matmul_index(gate_indices, number_of_qubits):
+    """Return the index string for Numpy.eignsum matrix multiplication.
+
+    The returned indices are to perform a matrix multiplication A.B where
+    the matrix A is an M-qubit matrix, matrix B is an N-qubit matrix, and
+    M <= N, and identity matrices are implied on the subsystems where A has no
+    support on B.
+
+    Args:
+        gate_indices (list[int]): the indices of the right matrix subsystems
+                                   to contract with the left matrix.
+        number_of_qubits (int): the total number of qubits for the right matrix.
+
+    Returns:
+        str: An indices string for the Numpy.einsum function.
+
+    Raises:
+        QISKitError: if the total number of qubits plus the number of
+        contracted indices is greater than 26.
+    """
+
+    # Since we use ASCII alphabet for einsum index labels we are limited
+    # to 26 total free left (lowercase) and 26 right (uppercase) indexes.
+    # The rank of the contracted tensor reduces this as we need to use that
+    # many characters for the contracted indices
+    if len(gate_indices) + number_of_qubits > 26:
+        raise QISKitError("Total number of free indexes limited to 26")
+
+    # Right indices for the N-qubit input and output tensor
+    idx_right = ascii_uppercase[:number_of_qubits]
+
+    # Left ndicies for N-qubit input tensor
+    idx_left_in = ascii_lowercase[:number_of_qubits]
+
+    # Left indices for the N-qubit output tensor
+    idx_left_out = list(idx_left_in)
+
+    # Left and right indices for the M-qubit multiplying tensor
+    mat_left = ""
+    mat_right = ""
+
+    # Update left indices for mat and output
+    for pos, idx in enumerate(reversed(gate_indices)):
+        mat_left += ascii_lowercase[-1 - pos]
+        mat_right += idx_left_in[-1 - idx]
+        idx_left_out[-1 - idx] = ascii_lowercase[-1 - pos]
+    idx_left_out = "".join(idx_left_out)
+
+    # Combine indices into matrix multiplication string format
+    # for numpy.einsum function
+    return "{mat_l}{mat_r}, ".format(mat_l=mat_left, mat_r=mat_right) + \
+           "{tens_lin}{tens_r}->{tens_lout}{tens_r}".format(tens_lin=idx_left_in,
+                                                            tens_lout=idx_left_out,
+                                                            tens_r=idx_right)
 
 
 # Functions used by the sympy simulators.