<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

Fix #894

### Summary

This PR changes the local unitary simulator to use the Numpy.einsum function. This removes the need to enlarge single and two qubit gates to full 2^N x 2^N matrices, and greatly speeds up performance for larger circuits.

### Details and comments

An example of runtime for a random circuit of 100 gates on my laptop (Macbook pro 3.1 GHz i7, 16 GB Ram):

Timing results for new implementation are:

```
2 qubit circuit, 100 gates, time taken: 0.073 s
3 qubit circuit, 100 gates, time taken: 0.072 s
4 qubit circuit, 100 gates, time taken: 0.071 s
5 qubit circuit, 100 gates, time taken: 0.075 s
6 qubit circuit, 100 gates, time taken: 0.087 s
7 qubit circuit, 100 gates, time taken: 0.08 s
8 qubit circuit, 100 gates, time taken: 0.11 s
9 qubit circuit, 100 gates, time taken: 0.352 s
10 qubit circuit, 100 gates, time taken: 1.346 s
11 qubit circuit, 100 gates, time taken: 4.931 s
12 qubit circuit, 100 gates, time taken: 28.891 s
```

Timing results for old implementation are

```
2 qubit circuit, 100 gates, time taken: 0.24 s
3 qubit circuit, 100 gates, time taken: 0.123 s
4 qubit circuit, 100 gates, time taken: 0.178 s
5 qubit circuit, 100 gates, time taken: 0.114 s
6 qubit circuit, 100 gates, time taken: 0.138 s
7 qubit circuit, 100 gates, time taken: 0.238 s
8 qubit circuit, 100 gates, time taken: 0.486 s
9 qubit circuit, 100 gates, time taken: 2.835 s
10 qubit circuit, 100 gates, time taken: 15.714 s
11 qubit circuit, 100 gates, time taken: 115.502 s
12 qubit circuit, 100 gates, time taken: ?
```
(I wasn't patient enough for the 12 qubit circuit)

The python file used for printing benchmarks was:

```python
# Simple benchmark
import numpy as np
import time
import qiskit

# Create a test circuit
for num_qubits in range(2, 13):
    # Create random test circuit
    num_gates = 100
    qr = qiskit.QuantumRegister(num_qubits)
    circ = qiskit.QuantumCircuit(qr)
    for j in range(num_gates):
        q = np.random.randint(0, num_qubits)
        r = np.random.randint(0, 4)
        if r == 0:       
            circ.cx(qr[q], qr[(q + 1) % num_qubits])
        elif r == 1:
            circ.h(qr[q])
        elif r == 2:
            circ.s(qr[q])
        elif r == 3:
            circ.t(qr[q])
    # Execute with timer
    start = time.time()
    result = qiskit.execute(circ, 'local_unitary_simulator').result()
    stop = time.time()
    print("{} qubit circuit, {} gates, time taken: {} s".format(num_qubits, num_gates, np.round(stop-start,3)))
```