@delapuente 2018-09-13
Remember to open an issue first and link it with "Fix <issuenumber>". I did it for you this time.