@ajavadia 2018-09-13
This looks great. It was really slow to enlarge every small gate before. I also like the revamp of tests. Once the linter passes, I think it can be merged.