@jaygambetta 2018-09-13
I agree